alx_travel_app/listings/requirements.txt
# Core Django & REST Framework
Django>=5.2,<6.0
djangorestframework>=3.15
drf-nested-routers>=0.93.3
django-filter>=24.1

# Environment variables management
python-decouple>=3.8
django-environ>=0.11.2

# Celery & Redis (for async tasks like payment emails)
celery>=5.3
redis>=5.0

# Email handling
django-celery-email>=3.0.0  # optional, if you want Celery for email backend

# For handling payments (Chapa)
requests>=2.32
